+++ dev/src/main.ts
@@ -6,6 +6,7 @@ import {
 	generateLong,
 	generatePrimary,
 	generateSquare,
+	generateShort,
 } from '../../src/signal-flags';
 
 const createEl = (html: string, el = 'td') => {
@@ -24,10 +25,11 @@ const createEl = (html: string, el = 'td') => {
 	document.body.prepend(div);
 }
 
-const svg = generateDefault().svg;
+const svgDefault = generateDefault().svg;
 const svgLong = generateLong().svg;
 const svgSquare = generateSquare().svg;
 const svgPrimary = generatePrimary().svg;
+const svgShort = generateShort().svg;
 
 const $app = document.querySelector<HTMLDivElement>('#app');
 
@@ -44,11 +46,15 @@ $thead.append($tr);
 $tr.append(createEl('Key', 'th'));
 $tr.append(createEl('slug', 'th'));
 $tr.append(createEl('Category', 'th'));
-$tr.append(createEl('Default', 'th'));
-$tr.append(createEl('Square', 'th'));
-$tr.append(createEl('Long', 'th'));
+$tr.append(createEl('Default 360 (pennants 540, card 180)', 'th'));
+$tr.append(
+	createEl('Square 240 (pennants 540, triangles 360, card 180)', 'th'),
+);
+$tr.append(createEl('Short (320, pennants 360, triangles 320, card 180', 'th'));
+$tr.append(createEl('Long 720', 'th'));
 $tr.append(createEl('Primary', 'th'));
 
+const scale = 3 / 100;
 for (const [key, flag] of Object.entries(flags)) {
 	const $tr = document.createElement('tr');
 	$tbody.append($tr);
@@ -57,20 +63,33 @@ for (const [key, flag] of Object.entries(flags)) {
 
 	$tr.append(createEl(flag.category));
 
-	let div = createEl(svg[key], 'div');
+	let div = createEl(svgDefault[key], 'div');
 	let td = createEl('');
 	td.append(div);
-	div.style.width = flag.shape === 'pennant' ? '18vw' : '12vw';
+	div.style.width =
+		flag.shape === 'pennant' ? `${540 * scale}vw`
+		: flag.dimensions === 'card' ? `${180 * scale}vw`
+		: `${360 * scale}vw`;
 	$tr.append(td);
 
 	div = createEl(svgSquare[key], 'div');
 	td = createEl('');
 	td.append(div);
 	div.style.width =
-		flag.shape === 'pennant' ? '18vw'
-		: flag.shape === 'triangle' ? '12vw'
-		: '8vw';
+		flag.shape === 'pennant' ? `${540 * scale}vw`
+		: flag.shape === 'triangle' ? `${360 * scale}vw`
+		: flag.dimensions === 'card' ? `${180 * scale}vw`
+		: `${240 * scale}vw`;
+	$tr.append(td);
 
+	div = createEl(svgShort[key], 'div');
+	td = createEl('');
+	td.append(div);
+	div.style.width =
+		flag.shape === 'pennant' ? `${360 * scale}vw`
+		: flag.shape === 'triangle' ? `${320 * scale}vw`
+		: flag.dimensions === 'card' ? `${180 * scale}vw`
+		: `${320 * scale}vw`;
 	$tr.append(td);
 
 	div = createEl(svgLong[key], 'div');