+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
-## [4.0.0](https://github.com/signal-flags/signal-flags-ts/releases/) - _2025-09-15_
+## [4.1.0](https://github.com/signal-flags/signal-flags-ts/tree/v4.1.0) - _2025-09-16_
+
+### Design changes
+
+- add `short` option for dimensions (4:3 for rectangles, 2:1 for pennants and
+  4:3 for triangles).
+- make flags A and B the right dimensions in the `square` set.
+- make the cross thicker on non-square R and X flags.
+
+### API changes
+
+- `feat` add `generateShort` to support short option.
+
+## [4.0.0](https://github.com/signal-flags/signal-flags-ts/tree/v4.0.0) - _2025-09-15_
 
 v4 is a ground-up rewrite in Typescript with a completely new API.